# docs/swagger.yaml
basePath: /
definitions:
  models.Movie:
    properties:
      id:
        type: string
      language:
        type: string
      length:
        type: integer
      name:
        type: string
    required:
    - language
    - length
    - name
    type: object
  models.Seat:
    properties:
      id:
        type: string
      number:
        type: integer
      theater_id:
        type: string
    type: object
  models.Show:
    properties:
      MovieId:
        type: string
      StartTime:
        type: integer
      TheaterId:
        type: string
      id:
        type: string
    type: object
  models.Theater:
    properties:
      address:
        type: string
      id:
        type: string
      name:
        type: string
      seats:
        items:
          $ref: '#/definitions/models.Seat'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /movies:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
    post:
      parameters:
      - description: Movie JSON
        in: body
        name: Movie
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Movie'
  /movies/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
    patch:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
  /shows:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Show'
            type: array
    post:
      parameters:
      - description: show JSON
        in: body
        name: show
        required: true
        schema:
          $ref: '#/definitions/models.Show'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Show'
  /shows/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Show'
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Show'
    patch:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: show JSON
        in: body
        name: show
        required: true
        schema:
          $ref: '#/definitions/models.Show'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Show'
  /theaters:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Theater'
            type: array
    post:
      parameters:
      - description: Theater JSON
        in: body
        name: Theater
        required: true
        schema:
          $ref: '#/definitions/models.Theater'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Theater'
  /theaters/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Theater'
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Theater'
    patch:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Theater JSON
        in: body
        name: Theater
        required: true
        schema:
          $ref: '#/definitions/models.Theater'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Theater'
swagger: "2.0"